backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.23
pyrogram==2.0.106
python-dateutil==2.8.2
pydantic==2.4.2
alembic==1.12.1
